# PERF: Using Numpy C-API arange