@jbrockmendel 2020-03-13
Since the perf numbers are pretty similar, can you report a few rounds of each