@jbrockmendel 2020-03-13
thats a nice improvement for the small cases, thanks