@jbrockmendel 2020-03-13
lgtm, thanks for accomodating my nitpicking