@WillAyd 2020-03-13
Why does this generate anything different from Cython? Shouldn't it generate the same result?