@WillAyd 2020-03-13
So doesn’t work like the built in range call then huh? Was expecting it would automatically convert to C without the Python lookup

Sent from my iPhone

> On Mar 13, 2020, at 3:34 PM, jbrockmendel <notifications@github.com> wrote:
> 
> ﻿
> Why does this generate anything different from Cython? Shouldn't it generate the same result?
> 
> The generated C code has to call np.arange in python-space (and lookup np.int64, and ...)
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub, or unsubscribe.